refactor(FakerError): rename file
