+++ src/finance.ts
@@ -1,7 +1,7 @@
 import type { Faker } from '.';
 import type { Helpers } from './helpers';
 import ibanLib from './iban';
-import { FakerError } from './internal/faker-js.error';
+import { FakerError } from './internal/faker.error';
 
 /**
  * Module to generate finance related entries.
