+++ test/mersenne.spec.ts
@@ -1,5 +1,5 @@
 import { beforeAll, beforeEach, describe, expect, it } from 'vitest';
-import { FakerError } from '../src/internal/faker-js.error';
+import { FakerError } from '../src/internal/faker.error';
 import { Mersenne } from '../src/mersenne';
 
 type SeededRun = {
